read handlers were outgrowing a single python file, which is a good thing: they outgrew because they got more mature exception handling and cleaner helper functions-- so i split them into separate files
